Переделка бота на несколько серверов

+++ src/main_bot/database.py
@@ -20,63 +20,26 @@ def job_database(func):
             if conn:
                 conn.close()
         return result
-
     return wrapper
 
 
-# Создаем таблицу
+
+# Создаем таблицы для проекта
 @job_database
-def create_table(cursor):
+def create_all_tables(cursor):
+    cursor.execute('''CREATE TABLE IF NOT EXISTS servers
+                      (id INTEGER PRIMARY KEY AUTOINCREMENT,
+                      name TEXT NOT NULL,
+                      domain TEXT NULL,
+                      ip TEXT NULL,
+                      main BOOLEAN default false
+                      )''')
     cursor.execute('''CREATE TABLE IF NOT EXISTS users
                     (id INTEGER PRIMARY KEY AUTOINCREMENT,
                     username TEXT NOT NULL,
                     password TEXT NOT NULL,
-                    role TEXT NOT NULL,
-                    status INTEGER NOT NULL default 1)''')
-    return 'edit'
-
-
-# Добавляем нового пользователя
-@job_database
-def add_user(cursor, username, password, role):
-    cursor.execute("INSERT INTO users (username, password, role) VALUES (?, ?, ?)", (username, password, role))
-    return 'edit'
-
-@job_database
-def delete_vpn_user(cursor, username):
-    cursor.execute("delete from users where username = ?",
-                   (username, ))
-    return 'edit'
-
-# Получить роль пользователя
-@job_database
-def get_role_user(cursor, username, password=None):
-    if password:
-        cursor.execute("SELECT role FROM users WHERE username=? and password=?", (username, password))
-    else:
-        cursor.execute("SELECT role FROM users WHERE username=?", (username,))
-    result = cursor.fetchone()
-    return result[0] if result else None
-
-
-# Получить список пользователей
-@job_database
-def get_all_username(cursor):
-    cursor.execute("SELECT username, password FROM users")
-    users = cursor.fetchall()
-    return users
-
-
-@job_database
-def get_count_users(cursor):
-    cursor.execute("SELECT count(1) FROM users")
-    users = cursor.fetchone()
-    return users[0]
-
-
-# ЛОКАЛЬНАЯ ВСТАВКА, у некоторых функций меняется функционал
-@job_database
-def create_table_bot(cursor):
+                    status INTEGER NOT NULL default 1,
+                    server INTEGER REFERENCES servers NOT NULL)''')
     cursor.execute('''CREATE TABLE IF NOT EXISTS tg_users
                     (id INTEGER PRIMARY KEY AUTOINCREMENT,
                     telegram_user_id TEXT NOT NULL,
@@ -89,13 +52,42 @@ def create_table_bot(cursor):
     return 'edit'
 
 
+
+# Работа с пользователями из таблицы users
+@job_database
+def add_user(cursor, username, password, server):
+    cursor.execute("INSERT INTO users (username, password, server) VALUES (?, ?, ?)", (username, password, server))
+    return 'edit'
+
+@job_database
+def delete_vpn_user(cursor, username, server):
+    cursor.execute("delete from users where username = ?", (username, ))
+    return 'edit'
+
 @job_database
 def get_count_users(cursor):
-    cursor.execute("SELECT count(1) FROM tg_users")
+    cursor.execute("SELECT count(1) FROM users")
     users = cursor.fetchone()
     return users[0]
 
 
+@job_database
+def enable_user_vpn(cursor, username, enabled, server):
+    cursor.execute("update users set status = ? WHERE username=?", (enabled, username))
+    return 'edit'
+
+@job_database
+def get_all_username_vpn(cursor, server = None, enabled=None):
+    cursor.execute("SELECT username, password, status, server FROM users where coalesce(?, status) = status and coalesce(?, server) = server"
+                   , (enabled, server))
+    users = cursor.fetchall()
+    if not server is None:
+        users = [usr[:-1] for usr in users]
+    return users
+
+
+
+# Работа с пользователями TG из таблицы tg_users
 @job_database
 def add_tg_user(cursor, telegram_user_id, username, role, enabled):
     if username is None:
@@ -106,10 +98,8 @@ def add_tg_user(cursor, telegram_user_id, username, role, enabled):
 
 @job_database
 def delete_tg_user(cursor, telegram_user_id):
-    cursor.execute("delete from tg_users where telegram_user_id = ?",
-                   (telegram_user_id, ))
+    cursor.execute("delete from tg_users where telegram_user_id = ?", (telegram_user_id, ))
     return 'edit'
-
 
 @job_database
 def get_all_tg_username(cursor, username=None, enabled=None):
@@ -120,6 +110,11 @@ def get_all_tg_username(cursor, username=None, enabled=None):
     users = cursor.fetchall()
     return users
 
+@job_database
+def get_count_users(cursor):
+    cursor.execute("SELECT count(1) FROM tg_users")
+    users = cursor.fetchone()
+    return users[0]
 
 @job_database
 def get_enable_user(cursor, user_id_chat):
@@ -127,13 +122,11 @@ def get_enable_user(cursor, user_id_chat):
     result = cursor.fetchone()
     return result[0] if result else None
 
-
 @job_database
 def get_role_user(cursor, user_id_chat):
     cursor.execute("SELECT role FROM tg_users WHERE telegram_user_id=?", (user_id_chat,))
     result = cursor.fetchone()
     return result[0] if result else None
-
 
 @job_database
 def enable_user_tg(cursor, username, enabled):
@@ -141,11 +134,9 @@ def enable_user_tg(cursor, username, enabled):
     return 'edit'
 
 
-@job_database
-def enable_user_vpn(cursor, username, enabled):
-    cursor.execute("update users set status = ? WHERE username=?", (enabled, username))
-    return 'edit'
 
+
+# Работа с таблицей связи пользователя TG с его аккаунтами tg_ls_usr_vpn
 
 @job_database
 def get_my_account(cursor, user_id):
@@ -157,7 +148,6 @@ def get_my_account(cursor, user_id):
     users = cursor.fetchall()
     return users
 
-
 @job_database
 def connect_tg_vpn(cursor, usr_id, vpn_name):
     cursor.execute(
@@ -165,7 +155,6 @@ def connect_tg_vpn(cursor, usr_id, vpn_name):
         values ((select id from tg_users where telegram_user_id = ?), (select id from users where username = ?))''',
         (usr_id, vpn_name))
     return 'edit'
-
 
 @job_database
 def unconnect_user(cursor, usr_id, vpn_name):
@@ -177,8 +166,27 @@ def unconnect_user(cursor, usr_id, vpn_name):
     return 'edit'
 
 
+
+# Информация по серверам из таблицы servers
 @job_database
-def get_all_username_vpn(cursor, enabled=None):
-    cursor.execute("SELECT username, password, status FROM users where (status = ? or ? is null)", (enabled, enabled))
+def add_server(cursor, name, domain, ip, main = False):
+    cursor.execute("INSERT INTO servers (name, domain, ip, main) VALUES (?, ?, ?, ?)", (name, domain, ip, main))
+    return 'edit'
+
+@job_database
+def delete_server(cursor, id_server):
+    cursor.execute("delete from servers where id = ?", (id_server, ))
+    return 'edit'
+
+@job_database
+def get_count_servers(cursor):
+    cursor.execute("SELECT count(1) FROM servers")
+    users = cursor.fetchone()
+    return users[0]
+
+
+@job_database
+def get_all_servers(cursor, enabled=None):
+    cursor.execute("SELECT id, name, domain, ip, main FROM servers")
     users = cursor.fetchall()
     return users
